requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
django==4.2.21
django-filter==24.3
djangorestframework==3.15.2
kombu==5.5.3
prompt-toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
gunicorn
celery
